requirements.txt
python-binance==1.0.17
python-telegram-bot==13.14
python-dotenv==0.19.2
pandas==1.4.2
numpy==1.22.3
Flask==2.2.5
Werkzeug==2.2.3
flask-login==0.6.2
sqlalchemy==1.4.36
pymysql==1.0.2
cryptography==37.0.2
websocket-client==1.3.2
ta==0.10.2
plotly==5.7.0
dash==2.3.1
redis==4.2.2
schedule==1.1.0
aiohttp==3.8.1
asyncio==3.4.3
